standard name change
changed this standard's name from Symfony2 to a more generic Symfony to prevent future confusion

fixes #65
