+++ Symfony/Sniffs/Formatting/BlankLineBeforeReturnSniff.php
@@ -1,0 +1,98 @@
+<?php
+
+/**
+ * This file is part of the Symfony2-coding-standard (phpcs standard)
+ *
+ * PHP version 5
+ *
+ * @category PHP
+ * @package  Symfony2-coding-standard
+ * @author   Authors <Symfony2-coding-standard@escapestudios.github.com>
+ * @license  http://spdx.org/licenses/MIT MIT License
+ * @link     https://github.com/escapestudios/Symfony2-coding-standard
+ */
+
+/**
+ * Throws errors if there's no blank line before return statements.
+ * Symfony coding standard specifies: "Add a blank line before return statements,
+ * unless the return is alone inside a statement-group (like an if statement);"
+ *
+ * PHP version 5
+ *
+ * @category PHP
+ * @package  Symfony2-coding-standard
+ * @author   Authors <Symfony2-coding-standard@escapestudios.github.com>
+ * @license  http://spdx.org/licenses/MIT MIT License
+ * @link     https://github.com/escapestudios/Symfony2-coding-standard
+ */
+namespace Symfony\Sniffs\Formatting;
+
+use PHP_CodeSniffer\Sniffs\Sniff;
+use PHP_CodeSniffer\Files\File;
+
+class BlankLineBeforeReturnSniff implements Sniff
+{
+    /**
+     * A list of tokenizers this sniff supports.
+     *
+     * @var array
+     */
+    public $supportedTokenizers = array(
+        'PHP',
+        'JS',
+    );
+
+    /**
+     * Returns an array of tokens this test wants to listen for.
+     *
+     * @return array
+     */
+    public function register()
+    {
+        return array(T_RETURN);
+    }
+
+    /**
+     * Processes this test, when one of its tokens is encountered.
+     *
+     * @param File $phpcsFile All the tokens found in the document.
+     * @param int  $stackPtr  The position of the current token in
+     *                        the stack passed in $tokens.
+     *
+     * @return void
+     */
+    public function process(File $phpcsFile, $stackPtr)
+    {
+        $tokens          = $phpcsFile->getTokens();
+        $current         = $stackPtr;
+        $previousLine    = $tokens[$stackPtr]['line'] - 1;
+        $prevLineTokens  = array();
+
+        while ($current >= 0 && $tokens[$current]['line'] >= $previousLine) {
+            if ($tokens[$current]['line'] == $previousLine
+                && $tokens[$current]['type'] !== 'T_WHITESPACE'
+                && $tokens[$current]['type'] !== 'T_COMMENT'
+                && $tokens[$current]['type'] !== 'T_DOC_COMMENT_CLOSE_TAG'
+                && $tokens[$current]['type'] !== 'T_DOC_COMMENT_WHITESPACE'
+            ) {
+                $prevLineTokens[] = $tokens[$current]['type'];
+            }
+            $current--;
+        }
+
+        if (isset($prevLineTokens[0])
+            && ($prevLineTokens[0] === 'T_OPEN_CURLY_BRACKET'
+            || $prevLineTokens[0] === 'T_COLON')
+        ) {
+            return;
+        } else if (count($prevLineTokens) > 0) {
+            $phpcsFile->addError(
+                'Missing blank line before return statement',
+                $stackPtr,
+                'Invalid'
+            );
+        }
+
+        return;
+    }
+}
