+++ Symfony/Tests/Formatting/BlankLineBeforeReturnUnitTest.inc
