--- Symfony2/Sniffs/WhiteSpace/BinaryOperatorSpacingSniff.php
@@ -1,79 +1,0 @@
-<?php
-
-/**
- * This file is part of the Symfony2-coding-standard (phpcs standard)
- *
- * PHP version 5
- *
- * @category PHP
- * @package  Symfony2-coding-standard
- * @author   Authors <Symfony2-coding-standard@escapestudios.github.com>
- * @license  http://spdx.org/licenses/MIT MIT License
- * @link     https://github.com/escapestudios/Symfony2-coding-standard
- */
-
-namespace Symfony2\Sniffs\Whitespace;
-
-use PHP_CodeSniffer\Sniffs\Sniff;
-use PHP_CodeSniffer\Files\File;
-use PHP_CodeSniffer\Util\Tokens;
-
-/**
- * SBinaryOperatorSpacingSniff.
- *
- * Throws warnings if a binary operator isn't surrounded with whitespace.
- *
- * PHP version 5
- *
- * @category PHP
- * @package  Symfony2-coding-standard
- * @author   Authors <Symfony2-coding-standard@escapestudios.github.com>
- * @license  http://spdx.org/licenses/MIT MIT License
- * @link     https://github.com/escapestudios/Symfony2-coding-standard
- */
-class SBinaryOperatorSpacingSniff implements Sniff
-{
-    /**
-     * A list of tokenizers this sniff supports.
-     *
-     * @var array
-     */
-    public $supportedTokenizers = array(
-                                   'PHP',
-                                  );
-
-    /**
-     * Returns an array of tokens this test wants to listen for.
-     *
-     * @return array
-     */
-    public function register()
-    {
-        return Tokens::$comparisonTokens;
-
-    }
-
-    /**
-     * Processes this test, when one of its tokens is encountered.
-     *
-     * @param File $phpcsFile The file being scanned.
-     * @param int  $stackPtr  The position of the current token
-     *                        in the stack passed in $tokens.
-     *
-     * @return void
-     */
-    public function process(File $phpcsFile, $stackPtr)
-    {
-        $tokens = $phpcsFile->getTokens();
-
-        if ($tokens[$stackPtr -1]['code'] !== T_WHITESPACE
-            || $tokens[$stackPtr +1]['code'] !== T_WHITESPACE
-        ) {
-            $phpcsFile->addError(
-                'Add a single space around binary operators',
-                $stackPtr,
-                'Invalid'
-            );
-        }
-    }
-}
