+++ build.properties
@@ -5,4 +5,4 @@ composer.path = ${basedir}/composer.phar
 #PHP Code Sniffer
 phpcs.standard = PEAR
 phpcs.dir = ${basedir}/vendor/squizlabs/php_codesniffer
-phpcs.symlink.path = vendor/squizlabs/php_codesniffer/src/Standards/Symfony2
+phpcs.symlink.path = vendor/squizlabs/php_codesniffer/src/Standards/Symfony
