+++ build.xml
@@ -1,4 +1,4 @@
-<project name="Symfony2-coding-standard" default="test">
+<project name="Symfony-coding-standard" default="test">
     <property file="build.properties" />
     <available property="composer.exists" file="${composer.path}" />
     <available property="symlink.exists" file="${phpcs.symlink.path}" />
@@ -36,7 +36,7 @@
     </target>
 
     <target name="symlink-this-coding-standard" description="Symlink this coding standard" unless="symlink.exists">
-        <symlink link="${basedir}/${phpcs.symlink.path}" resource="${basedir}/Symfony2"/>
+        <symlink link="${basedir}/${phpcs.symlink.path}" resource="${basedir}/Symfony"/>
     </target>
 
     <target name="phpunit" depends="symlink-this-coding-standard" description="Run unit tests with PHPUnit">
